requirements.txt
python-telegram-bot[all]>=20.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic-settings>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
